12/06/2025

+++ controller/deactivateController.php
@@ -12,44 +12,80 @@ if (!isset($_GET['uid'])) {
 // cast to integer cleans the input into a number and ignores anything that isn't a number.
 $userID = (int) $_GET['uid']; // Cast to integer
 
+if (!isset($_GET['action']) || !in_array($_GET['action'], ['deactivate', 'activate'])) {
+    $_SESSION['status_message'] = "Invalid action.";
+    header("Location: userlist");
+    exit();
+}
+
+$action = $_GET['action'];
+
+
 
 // delete all comments by selected user
 
-$cdelete = $conn->prepare("DELETE FROM `comment` WHERE `user` = ?;");
-
-if ($cdelete) {
-    $cdelete->bind_param("i", $userID);
-
-    if ($cdelete->execute()) {
-        $_SESSION['status_message'] = "Comments deleted successfully!";
-    } else {
-        $_SESSION['status_message'] = "Error executing query: " . $conn->error;
+if ($action == 'deactivate') {
+    // delete all comments by selected user
+    $cdelete = $conn->prepare("DELETE FROM `comment` WHERE `user` = ?;");
+    if ($cdelete) {
+        $cdelete->bind_param("i", $userID);
+        $cdelete->execute();
+        $cdelete->close();
     }
 
-    $cdelete->close();
-} else {
-    $_SESSION['status_message'] = "Error preparing query: " . $conn->error;
-}
+// $cdelete = $conn->prepare("DELETE FROM `comment` WHERE `user` = ?;");
+
+// if ($cdelete) {
+//     $cdelete->bind_param("i", $userID);
+
+//     if ($cdelete->execute()) {
+//         $_SESSION['status_message'] = "Comments deleted successfully!";
+//     } else {
+//         $_SESSION['status_message'] = "Error executing query: " . $conn->error;
+//     }
+
+//     $cdelete->close();
+// } else {
+//     $_SESSION['status_message'] = "Error preparing query: " . $conn->error;
+// }
 
 // deactivate selected user
 
+$update = $conn->prepare("UPDATE `user` SET `status`='inactive' WHERE `id` = ?;");
+    if ($update) {
+        $update->bind_param("i", $userID);
+        $update->execute();
+        $update->close();
+    }
+    $_SESSION['status_message'] = "User deactivated successfully!";
+} elseif ($action == 'activate') {
+    // activate selected user (set status back to 'user')
+    $update = $conn->prepare("UPDATE `user` SET `status`='user' WHERE `id` = ?;");
+    if ($update) {
+        $update->bind_param("i", $userID);
+        $update->execute();
+        $update->close();
+    }
+    $_SESSION['status_message'] = "User activated successfully!";
+}
+
 // Prepare and execute statement
 // Using prepared statements will help prevent sql injection
-$delete = $conn->prepare("UPDATE `user` SET `status`='inactive' WHERE `id` = ?;");
+// $delete = $conn->prepare("UPDATE `user` SET `status`='inactive' WHERE `id` = ?;");
 
-if ($delete) {
-    $delete->bind_param("i", $userID);
+// if ($delete) {
+//     $delete->bind_param("i", $userID);
 
-    if ($delete->execute()) {
-        $_SESSION['status_message'] = "User deactivated successfully!";
-    } else {
-        $_SESSION['status_message'] = "Error executing query: " . $conn->error;
-    }
+//     if ($delete->execute()) {
+//         $_SESSION['status_message'] = "User deactivated successfully!";
+//     } else {
+//         $_SESSION['status_message'] = "Error executing query: " . $conn->error;
+//     }
 
-    $delete->close();
-} else {
-    $_SESSION['status_message'] = "Error preparing query: " . $conn->error;
-}
+//     $delete->close();
+// } else {
+//     $_SESSION['status_message'] = "Error preparing query: " . $conn->error;
+// }
 
 // Redirect back to the user list page
 header("Location: userlist");
